fla/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fla"
version = "0.1.0"
description = "Fine-tune VLA (Vision-Language-Action) models with state-of-the-art methods"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    {name = "FLA Contributors"}
]
keywords = ["robotics", "vision-language-action", "fine-tuning", "pi0", "lerobot"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core ML
    "jax[cuda12]>=0.4.30",
    "flax>=0.8.0",
    "optax>=0.2.0",
    "orbax-checkpoint>=0.5.0",

    # Data
    "lerobot>=0.1.0",
    "torch>=2.0.0",
    "numpy>=1.24.0",

    # Image processing
    "augmax>=0.3.0",
    "Pillow>=10.0.0",

    # Utilities
    "einops>=0.7.0",
    "sentencepiece>=0.2.0",
    "tqdm>=4.65.0",
    "tyro>=0.7.0",
    "safetensors>=0.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
eval = [
    "gymnasium>=0.29.0",
    "gym-aloha>=0.1.0",
    "mujoco>=3.0.0",
]

[project.scripts]
fla-train = "fla.scripts.train:main"
fla-eval = "fla.scripts.evaluate:main"
fla-serve = "fla.scripts.serve:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["fla*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
